+++ flow-typed/common/alert.js
@@ -1,2 +1,7 @@
 declare type AlertVariant = "raised" | "flat";
 declare type AlertColor = "success" | "info" | "error" | "warning";
+
+declare type AlertClasses = {
+  root?: string,
+  content?: string,
+};