+++ src/alert/alert.jsx
@@ -13,6 +13,13 @@ type Props = {
   message: Node,
   /** Classname applied to the alert component */
   className?: string,
+  /**
+   * Provides classNames to the alert and its sub-components. Options include:
+   *  - `root`
+   *
+   *  - `content`
+   */
+  classes?: AlertClasses,
   /** The color styling of the alert. Defaults to `error` */
   color?: AlertColor,
   /** When "center", aligns the Alert's messages horizontally */
@@ -33,16 +40,16 @@ const colorIcon = {
  */
 export const Alert = (props: Props) => {
   const {
-    textAlign = "", className = "", color = "error", message, variant = "", ...alertProps
+    textAlign = "", className = "", classes = {}, color = "error", message, variant = "", ...alertProps
   } = props;
   const Icon = colorIcon[color];
   return (
     <div
       data-testid="alert"
-      className={styles.alertContainer}
+      className={classNames(styles.alertContainer, classes.root)}
     >
       <SnackbarContent
-        className={classNames(styles[color], styles[variant], `alert-${color}`, className)}
+        className={classNames(styles[color], styles[variant], `alert-${color}`, className, classes.content)}
         classes={{
           message: styles[textAlign],
         }}