+++ styleguide.config.js
@@ -135,5 +135,11 @@ module.exports = {
         },
       ],
     },
+    resolve: {
+      alias: {
+        "@grail/components": path.resolve(__dirname, "src"),
+        "@grail/lib": path.resolve(__dirname, "../lib/src"),
+      },
+    },
   },
 };